Refactores GameEventListeners to allow an easier extension
