+++ Assets/FeTo/SO_Architecture/GameEvents/Base/GameEventListener.cs
@@ -5,30 +5,20 @@ using Logger = FeTo.Logging.Logger;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public abstract class GameEventListener<T, Q> : MonoBehaviour where Q : UnityEvent<T>
+    public class GameEventListener : MonoBehaviour
     {
         [Tooltip("Event to register with.")]
-        public GameEvent<T, Q> Event;
+        public GameEvent Event;
 
         [Tooltip("Response to invoke when Event is raised.")]
-        public Q Response;
+        public UnityEvent Response;
 
-        protected void OnEnable()
-        {
-            Event.RegisterListener(this);
-        }
-
-        protected void OnDisable()
-        {
-            Event.UnregisterListener(this);
-        }
-
-        public void OnEventRaised(T value)
+        public void OnEventRaised()
         {
 #if UNITY_EDITOR
             Logger.FeToInfo(Event.LogEvents, $"FeTo: Event {Event.name} catched by {gameObject.name}");
 #endif
-            Response.Invoke(value);
+            Response.Invoke();
         }
     }
 }
