+++ Assets/FeTo/SO_Architecture/GameEvents/Base/GameEventListenerTQ.cs
@@ -5,30 +5,20 @@ using Logger = FeTo.Logging.Logger;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public class GameEventListener : MonoBehaviour
+    public abstract class GameEventListener<T, Q> : MonoBehaviour where Q : UnityEvent<T>
     {
         [Tooltip("Event to register with.")]
-        public GameEvent Event;
+        public GameEvent<T, Q> Event;
 
         [Tooltip("Response to invoke when Event is raised.")]
-        public UnityEvent Response;
+        public Q Response;
 
-        protected void OnEnable()
-        {
-            Event.RegisterListener(this);
-        }
-
-        protected void OnDisable()
-        {
-            Event.UnregisterListener(this);
-        }
-
-        public void OnEventRaised()
+        public void OnEventRaised(T value)
         {
 #if UNITY_EDITOR
             Logger.FeToInfo(Event.LogEvents, $"FeTo: Event {Event.name} catched by {gameObject.name}");
 #endif
-            Response.Invoke();
+            Response.Invoke(value);
         }
     }
 }
