+++ Assets/FeTo/SO_Architecture/GameEvents/OnEnableGameEventListeners/OnEnableBoolGameEventListener.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public class BoolGameEventListener : GameEventListener<bool, BoolEvent> { }
+    public class OnEnableBoolGameEventListener : OnEnableGameEventListener<bool, BoolEvent> { }
 
 }
