+++ Assets/FeTo/SO_Architecture/GameEvents/OnEnableGameEventListeners/OnEnableFloatGameEventListener.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public class FloatGameEventListener : GameEventListener<float, FloatEvent> { }
+    public class OnEnableFloatGameEventListener : OnEnableGameEventListener<float, FloatEvent> { }
 }
