+++ Assets/FeTo/SO_Architecture/GameEvents/OnEnableGameEventListeners/OnEnableGameEventListener.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+namespace FeTo.SOArchitecture
+{
+    [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
+    public class OnEnableGameEventListener : GameEventListener
+    {
+        protected void OnEnable()
+        {
+            Event.RegisterListener(this);
+        }
+
+        protected void OnDisable()
+        {
+            Event.UnregisterListener(this);
+        }
+    }
+}
