+++ Assets/FeTo/SO_Architecture/GameEvents/OnEnableGameEventListeners/OnEnableIntGameEventListener.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public class IntGameEventListener : GameEventListener<int, IntEvent> { }
+    public class OnEnableIntGameEventListener : OnEnableGameEventListener<int, IntEvent> { }
 }
