+++ Assets/FeTo/SO_Architecture/GameEvents/OnEnableGameEventListeners/OnEnableStringGameEventListener.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 namespace FeTo.SOArchitecture
 {
     [HelpURL("https://github.com/Neengash/UnityFeTo/tree/FeTo/SO_Architecture/GameEvents#scriptable-object---gameevents")]
-    public class StringGameEventListener : GameEventListener<string, StringEvent> { }
+    public class OnEnableStringGameEventListener : OnEnableGameEventListener<string, StringEvent> { }
 }
